requirements.txt
asgiref==3.5.2
distlib==0.3.6
dj-database-url
Django==3.1.14
django-heroku==0.3.1
djangorestframework==3.14.0
filelock==3.8.0
gunicorn==20.1.0
platformdirs==2.5.2
psycopg2==2.9.5
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
validate-email==1.3
virtualenv==20.16.6
whitenoise==6.2.0
